fix(recovery): remove `claims` dependency
Prior to this change, the recovery module relied on the parameters
stored in the claims module to determine if an asset is recoverable (if
sent to the Cosmos address instead of the Ethereum address for a given
mnemonic). This change removes the dependency and allows the recovery
module to store this information thereby making it possible to remove
the claims module.